Use fast_structured_np_unique in gltf2_blender_mesh.merge_duplicate_verts()
merge_duplicate_verts() also uses np.unique on a structured array.

To be able to use fast_structured_np_unique from gltf2_blender_mesh,
fast_structured_np_unique has been moved to gltf2_blender_conversion.
